moved seperate tests into several files and added them to tests folder. moved examples into examples folder, introduced makefile
