+++ requirements.txt
@@ -1,6 +1,5 @@
 flask
 matplotlib
-#tensorflow
 graphitesend
 antlr4-python3-runtime
 numpy
