+++ tests/context.py
@@ -1,0 +1,14 @@
+import os
+import sys
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname("bitflow"), '..')))
+
+import bitflow.processingstep as processingstep
+import bitflow.script_parser as script_parser
+import bitflow.sinksteps as sinksteps
+import bitflow.batchprocessingstep as batchprocessingstep
+import bitflow.pipeline as bitflow_pipeline
+import bitflow.source as source
+import bitflow.batch as batch
+
+import bitflow.fork as fork
+import bitflow.helper as helper
